+++ src/main/java/at/fhv/e_commerce_application/rest/dtos/order/GetOrderItemDTO.java
@@ -1,11 +1,12 @@
 package at.fhv.e_commerce_application.rest.dtos.order;
 
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record GetOrderItemDTO(
         UUID id,
         UUID productId,
         int quantity,
-        double price
+        BigDecimal price
 ){}
