+++ src/main/java/at/fhv/e_commerce_application/rest/dtos/products/GetProductDTO.java
@@ -1,5 +1,6 @@
 package at.fhv.e_commerce_application.rest.dtos.products;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 
@@ -7,7 +8,7 @@ public record GetProductDTO(
         UUID id,
         String name,
         String description,
-        double price,
+        BigDecimal price,
         int stock,
         String status
 ) {}
